Move all the Dev Services related classes to new module
