+++ extensions/devservices/datasource-common/src/main/java/io/quarkus/devservices/datasource/common/DevServicesDatasourceConfigurationHandlerBuildItem.java
@@ -1,4 +1,4 @@
-package io.quarkus.datasource.deployment.spi;
+package io.quarkus.devservices.datasource.common;
 
 import java.util.HashMap;
 import java.util.List;
