+++ extensions/devservices/datasource-common/src/main/java/io/quarkus/devservices/datasource/common/DevServicesDatasourceResultBuildItem.java
@@ -1,4 +1,4 @@
-package io.quarkus.datasource.deployment.spi;
+package io.quarkus.devservices.datasource.common;
 
 import java.util.Collections;
 import java.util.Map;
