+++ extensions/devservices/h2/pom.xml
@@ -17,6 +17,10 @@
             <artifactId>quarkus-datasource-deployment-spi</artifactId>
         </dependency>
         <dependency>
+            <groupId>io.quarkus</groupId>
+            <artifactId>quarkus-devservices-datasource-common</artifactId>
+        </dependency>
+        <dependency>
             <groupId>com.h2database</groupId>
             <artifactId>h2</artifactId>
         </dependency>
